sua lai

+++ src/UI/JFquanLy_user.java
@@ -227,7 +227,7 @@ public void FillTable() {
     }//GEN-LAST:event_bntSuaActionPerformed
 
     private void bntXoaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_bntXoaActionPerformed
-               int[] selectedRows = tbBan.getSelectedRows();
+        int[] selectedRows = tbBan.getSelectedRows();
 
         if (selectedRows.length <= 0) {
             JOptionPane.showMessageDialog(null, "Bạn chưa tài khoản !");
@@ -240,32 +240,21 @@ public void FillTable() {
                 String tenban = (String) tbBan.getValueAt(i, 1);
 
                 sp += tenban + "\n";
-            }
-//            int qs;
-//            qs = JOptionPane.showConfirmDialog(null, "Xóa tài khoản: \n " + sp, "Xóa tài khoản", JOptionPane.YES_NO_OPTION);
-//            if (qs == JOptionPane.YES_OPTION) {
-//                boolean xoa = dao.DeleteTaiKhoan(ListMaBan);
-//
-//                if (xoa == true) {
-//                    FillTable();
-// 
-//                }else
-//                JOptionPane.showMessageDialog(null, "Không xóa được tài khoản !");
-//
-//            }
-//        }
-//            
-        int nutbam = JOptionPane.showConfirmDialog(new JFrame(), "bạn chắc chắn xóa?", "xóA", JOptionPane.YES_NO_OPTION);
-        if (nutbam == JOptionPane.YES_OPTION) {
-            int cacdong[] = tbBan.getSelectedRows();
-            for (int i = 0; i < cacdong.length; i++) {
-                String MaPhong = tbBan.getValueAt(cacdong[i], 0).toString();
-                    dao.delete(MaPhong);
-                    FillTable();
-                
+            } 
+            int nutbam = JOptionPane.showConfirmDialog(new JFrame(), "bạn chắc chắn xóa?", "xóA", JOptionPane.YES_NO_OPTION);
+            if (nutbam == JOptionPane.YES_OPTION) {
+                int cacdong[] = tbBan.getSelectedRows();
+                for (int i = 0; i < cacdong.length; i++) {
+                    String MaPhong = tbBan.getValueAt(cacdong[i], 0).toString();
+                    if (MaPhong.matches(RUN.nv.getID() + "")) {
+                        JOptionPane.showMessageDialog(null, "Bạn không thể xoá chính bạn!");
+                    } else {
+                        dao.delete(MaPhong);
+                        FillTable();
 
+                    }
+                }
             }
-        }
         }
     }//GEN-LAST:event_bntXoaActionPerformed
